Create load.py file, mode read
