+++ hub/__init__.py
@@ -1,4 +1,4 @@
 from .api.dataset import Dataset
-from .core.tensor import read
+from .util.load import read
 
 __version__ = "2.0.0"
